+++ src/descriptor_table.rs
@@ -3,8 +3,8 @@
 use crate::addresses::*;
 use crate::asm;
 
-const LIMIT_INTERRUPT_DESCRIPTOR_TABLE: u32 = 0x000007FF;
-const LIMIT_GDT: u32 = 8 * 3 - 1;
+const LIMIT_INTERRUPT_DESCRIPTOR_TABLE: u16 = 0x000007FF;
+const LIMIT_GDT: u16 = 8 * 3 - 1;
 const ACCESS_RIGHT_IDT: u32 = 0x008E;
 
 #[repr(C, packed)]
@@ -87,7 +87,7 @@ fn init_idt() -> () {
     let interrupt_descriptor_table: *mut GateDescriptor =
         VIRTUAL_ADDRESS_IDT as *mut GateDescriptor;
 
-    const SIZE_IDT_ENTRY: u32 = 16;
+    const SIZE_IDT_ENTRY: u16 = 16;
     for i in 0..=(LIMIT_INTERRUPT_DESCRIPTOR_TABLE / SIZE_IDT_ENTRY) {
         unsafe {
             (*interrupt_descriptor_table.offset(i as isize)).set_gate_descriptor(0, 0, 0);
@@ -103,7 +103,7 @@ fn init_idt() -> () {
 fn init_gdt() -> () {
     let gdt: *mut SegmentDescriptor = VIRTUAL_ADDRESS_GDT as *mut SegmentDescriptor;
 
-    const SIZE_GDT_ENTRY: u32 = 8;
+    const SIZE_GDT_ENTRY: u16 = 8;
     for i in 0..=((LIMIT_GDT + 1) / SIZE_GDT_ENTRY) {
         unsafe {
             (*gdt.offset(i as isize)).set_segment_descriptor(SegmentType::NullSegment);
@@ -115,7 +115,7 @@ fn init_gdt() -> () {
         (*gdt.offset(2)).set_segment_descriptor(SegmentType::CodeSegment);
     }
 
-    asm::lgdt(LIMIT_GDT as u16, VIRTUAL_ADDRESS_GDT);
+    asm::lgdt(LIMIT_GDT, VIRTUAL_ADDRESS_GDT);
 }
 
 fn set_interruption() {
